Extract composer diff UI boundary
