+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.22",
-  "tag": "v3.4.22",
+  "version": "3.4.23",
+  "tag": "v3.4.23",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.21 <3.4.22"
+      ">=3.4.22 <3.4.23"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.21 first."
+    "message": "Update to v3.4.22 first."
   }
 }
